+++ py/sdrl/constants.py
@@ -7,6 +7,7 @@ HTML_DIFFICULTY_SIGN = "&#x26ab;&#xfe0e;"  # &#x26ab; is an icon and always blac
   # https://commons.wikimedia.org/wiki/Unicode_circle_shaped_symbols
 INTERACT_ACCEPT_SYMBOL = "✓"
 INTERACT_REJECT_SYMBOL = "X"
+HTACCESS_FILE = ".htaccess"  # in instructor part of build directory
 METADATA_FILE = "course.json"  # at top-level of build directory
 PARTICIPANT_FILE = "student.yaml"
 REPOS_HOME_VAR = "SEDRILA_INSTRUCTOR_REPOS_HOME"